Command processing & crash log
Currently the handler will process a small list of commmands to
demonstrate a proof of concept.

Additioanly, the sever will now log any and all exceptions thrown in the
context of the server itself. However the exception will still crash the
server.

I am working on a few ideas to fix the server hangup after the crash is
logged