Basic Transaction Wrapper
Added a simple transaction wrapper that will receivec requests from the
user, passed by the server. The wrapper will attempt to parse the
request, and return the relevent object data. Currently runs the
database, but does not actually retrieve the data from it.

Need to add flags to simpleDB.pl to allow for compatability with
Transaction Wrapper!